requirements.txt
Flask==2.1.3
Flask_JWT_Extended==4.4.2
Flask_Login==0.6.1
Flask_SQLAlchemy==2.5.1
pydantic==1.9.1
python-dotenv==0.20.0
python_magic==0.4.27
Werkzeug==2.1.2

python-magic~=0.4.27
SQLAlchemy~=1.4.39